added http pkg and changed chunk manager to an interface
